+++ coupon-server/src/main/java/com/drf/coupon/model/response/internal/InternalCartCouponCalculateResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record InternalCartCouponCalculateResponse(
         boolean applicable,
-        int totalDiscountAmount,
+        long totalDiscountAmount,
         List<InternalCouponItemResult> items
 ) {
 }
