+++ coupon-server/src/main/java/com/drf/coupon/model/response/internal/ProductCouponCalculateResponse.java
@@ -2,6 +2,6 @@ package com.drf.coupon.model.response.internal;
 
 public record ProductCouponCalculateResponse(
         boolean applicable,
-        int discountAmount
+        long discountAmount
 ) {
 }
