+++ coupon-server/src/main/java/com/drf/coupon/model/response/internal/ProductCouponResult.java
@@ -3,7 +3,7 @@ package com.drf.coupon.model.response.internal;
 public record ProductCouponResult(
         long memberCouponId,
         String name,
-        int discountAmount,
+        long discountAmount,
         boolean isBest,
         boolean usedOnOtherItem
 ) {
